+++ Source/Libraries/CorruptCore/BlastLayer.cs
